requirements.txt
streamlit==1.37.0
pandas
google-adk
google-cloud-aiplatform
vertexai
python-dotenv
